requirements.txt
beautifulsoup4==4.12.2
pandas==2.1.4
requests==2.31.0
selenium==4.18.1
tqdm==4.66.4
Pillow==10.4.0
scikit-learn==1.5.2
torch==2.5.1
torchvision==0.20.1
fastapi==0.115.5
uvicorn==0.32.0
